requirements/base.txt
amqp==5.3.1
asgiref==3.11.1
attrs==26.1.0
autobahn==25.12.2
Automat==25.4.16
billiard==4.2.4
cbor2==5.9.0
celery==5.6.3
cffi==2.0.0
channels==4.3.2
channels_redis==4.3.0
click==8.3.2
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
constantly==23.10.4
cryptography==46.0.7
daphne==4.2.1
Django==5.0
django-flower==1.0.0
django-jinja==2.11.0
django-redis==6.0.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
gevent==26.4.0
greenlet==3.4.0
humanize==0.5.1
hyperlink==21.0.0
idna==3.11
Incremental==24.11.0
Jinja2==3.1.6
kombu==5.6.2
MarkupSafe==3.0.3
msgpack==1.1.2
packaging==26.0
pillow==12.1.1
plumbum==1.10.0
prompt_toolkit==3.0.52
py-ubjson==0.16.1
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
PyJWT==2.11.0
pyOpenSSL==26.0.0
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2026.1.post1
redis==7.4.0
rpyc==6.0.2
service-identity==24.2.0
six==1.17.0
sqlparse==0.5.5
Twisted==25.5.0
txaio==25.12.2
typing_extensions==4.15.0
tzdata==2026.1
tzlocal==5.3.1
ujson==5.12.0
vine==5.1.0
wcwidth==0.6.0
zope.event==6.1
zope.interface==8.2
